#45 add access check for delete board

+++ modules/projects/application/handler/DeleteBoardHandler.php
@@ -3,6 +3,7 @@
 namespace modules\projects\application\handler;
 
 use modules\projects\application\command\DeleteBoardCommand;
+use modules\projects\application\port\IProjectAccess;
 use modules\projects\domain\repository\IBoardRepository;
 use modules\projects\domain\valueObject\BoardId;
 use RuntimeException;
@@ -10,10 +11,14 @@ use RuntimeException;
 class DeleteBoardHandler
 {
     private IBoardRepository $boardRepository;
+    private IProjectAccess $projectAccess;
 
-    public function __construct(IBoardRepository $boardRepository)
-    {
-        $this->boardRepository = $boardRepository;
+    public function __construct(
+        IBoardRepository $boardRepository,
+        IProjectAccess $projectAccess
+    ) {
+        $this->boardRepository  = $boardRepository;
+        $this->projectAccess    = $projectAccess;
     }
 
     public function handle(DeleteBoardCommand $command): void
@@ -25,7 +30,9 @@ class DeleteBoardHandler
             throw new RuntimeException("Board with ID {$command->id} not found");
         }
 
-        // TODO: проверка прав
+        if (!$this->projectAccess->canDeleteBoard($command->deletedBy, $command->id)) {
+            throw new RuntimeException('You are not allowed to delete this board');
+        }
 
         $this->boardRepository->remove($board);
     }
